MCR-3640 refactoring
