+++ mycore-webcli/src/main/vue/webcli/src/App.vue
@@ -5,35 +5,17 @@ import CommandHistoryDialog from '@/components/CommandHistoryDialog.vue';
 import CommandMenu from '@/components/CommandMenu.vue';
 import CommandSuggestions from '@/components/CommandSuggestions.vue';
 import SettingsDialog from '@/components/SettingsDialog.vue';
-import { useCommandHistory } from '@/composables/useCommandHistory';
-import { useCommandSearch } from '@/composables/useCommandSearch';
+import { useCommandInputController } from '@/composables/useCommandInputController';
+import { useSettingsState } from '@/composables/useSettingsState';
 import { useWebCliTransport } from '@/composables/useWebCliTransport';
-import { loadSettings, normalizeSettings, persistSettings } from '@/services/settings';
 
-const settings = ref(loadSettings());
-const command = ref('');
 const activeTab = ref<'log' | 'queue'>('log');
 const isSettingsDialogOpen = ref(false);
 const isCommandHistoryOpen = ref(false);
-const isSuggestionMenuVisible = ref(false);
-const isSuggestionNavigationActive = ref(false);
-const suppressSuggestionMenuOnFocus = ref(false);
 const isQueueExpanded = ref(false);
-const inputElement = ref<HTMLInputElement | null>(null);
-const executeButtonElement = ref<HTMLButtonElement | null>(null);
 const logElement = ref<HTMLElement | null>(null);
-const suggestionListId = 'webcli-command-suggestions';
 const queuePreviewSize = 99;
-
-const {
-  addEntry,
-  browseDown,
-  browseUp,
-  clear: clearCommandHistory,
-  entries: commandHistoryEntries,
-  finalizeExecution,
-  trimToSize,
-} = useCommandHistory(() => Math.max(0, settings.value.comHistorySize));
+const { settings } = useSettingsState();
 
 const {
   clearCommandQueue,
@@ -58,15 +40,32 @@ const {
 );
 
 const {
+  clearCommandHistory,
+  command,
+  commandHistoryEntries,
+  executeButtonElement,
+  executeCommand,
   hasSuggestions,
   highlightedIndex,
   highlightedSuggestion,
-  highlightSuggestion,
-  moveHighlight,
-  resetHighlight,
-  totalSuggestionCount,
+  inputElement,
+  isSuggestionMenuVisible,
+  onCommandInput,
+  onCommandInputBlur,
+  onCommandInputFocus,
+  onCommandKeydown,
+  onSuggestionActivate,
+  restoreCommandFromHistory,
+  selectCommand,
+  suggestionListId,
   suggestions,
-} = useCommandSearch(commandGroups, command, computed(() => settings.value.suggestionLimit));
+  totalSuggestionCount,
+} = useCommandInputController({
+  commandGroups,
+  commandHistorySize: computed(() => settings.value.comHistorySize),
+  suggestionLimit: computed(() => settings.value.suggestionLimit),
+  onExecuteCommand: runCommand,
+});
 
 const visibleQueue = computed(() => {
   return isQueueExpanded.value ? queue.value : queue.value.slice(0, queuePreviewSize);
@@ -82,168 +81,6 @@ const visibleLogLines = computed(() => {
   });
 });
 
-function areSettingsEqual(left: typeof settings.value, right: typeof settings.value): boolean {
-  return (
-    left.historySize === right.historySize &&
-    left.comHistorySize === right.comHistorySize &&
-    left.suggestionLimit === right.suggestionLimit &&
-    left.autoscroll === right.autoscroll &&
-    left.continueIfOneFails === right.continueIfOneFails
-  );
-}
-
-function closeSuggestionMenu(restoreFocus = false): void {
-  isSuggestionMenuVisible.value = false;
-  isSuggestionNavigationActive.value = false;
-  resetHighlight();
-  if (restoreFocus) {
-    suppressSuggestionMenuOnFocus.value = true;
-    nextTick(() => {
-      if (inputElement.value && document.activeElement !== inputElement.value) {
-        inputElement.value.focus();
-      }
-    });
-  }
-}
-
-function activateSuggestionNavigation(index = 0): void {
-  if (!hasSuggestions.value) {
-    return;
-  }
-  isSuggestionMenuVisible.value = true;
-  isSuggestionNavigationActive.value = true;
-  highlightSuggestion(index);
-}
-
-function selectNextPlaceholder(startFrom = 0): boolean {
-  const input = inputElement.value;
-  if (!input || !command.value) {
-    return false;
-  }
-  const regex = /\{[0-9]+\}/g;
-  regex.lastIndex = startFrom;
-  const match = regex.exec(command.value);
-  if (match) {
-    input.focus();
-    input.setSelectionRange(match.index, match.index + match[0].length);
-    return true;
-  }
-  return false;
-}
-
-function executeCommand(): void {
-  const value = command.value.trim();
-  if (!value) {
-    return;
-  }
-  finalizeExecution();
-  runCommand(value);
-  addEntry(value);
-  command.value = '';
-  closeSuggestionMenu();
-}
-
-function onCommandKeydown(event: KeyboardEvent): void {
-  if (event.key === 'ArrowDown' && isSuggestionMenuVisible.value && hasSuggestions.value) {
-    event.preventDefault();
-    if (!isSuggestionNavigationActive.value) {
-      activateSuggestionNavigation(Math.min(1, suggestions.value.length - 1));
-      return;
-    }
-    moveHighlight(1);
-    return;
-  }
-  if (event.key === 'ArrowUp' && isSuggestionMenuVisible.value && hasSuggestions.value) {
-    event.preventDefault();
-    if (!isSuggestionNavigationActive.value) {
-      activateSuggestionNavigation(suggestions.value.length - 1);
-      return;
-    }
-    moveHighlight(-1);
-    return;
-  }
-  if (event.key === 'Enter') {
-    if (isSuggestionMenuVisible.value && isSuggestionNavigationActive.value && highlightedSuggestion.value) {
-      event.preventDefault();
-      selectCommand(highlightedSuggestion.value.command);
-      return;
-    }
-    event.preventDefault();
-    executeCommand();
-    return;
-  }
-  if (event.key === 'Escape' && isSuggestionMenuVisible.value) {
-    event.preventDefault();
-    closeSuggestionMenu(true);
-    return;
-  }
-  if (event.key === 'ArrowUp') {
-    closeSuggestionMenu();
-    const previousCommand = browseUp(command.value);
-    if (previousCommand !== null) {
-      command.value = previousCommand;
-    }
-    return;
-  }
-  if (event.key === 'ArrowDown') {
-    closeSuggestionMenu();
-    const nextCommand = browseDown();
-    if (nextCommand !== null) {
-      command.value = nextCommand;
-    }
-    return;
-  }
-  if (event.key === 'Tab' && !event.shiftKey) {
-    if (
-      isSuggestionMenuVisible.value &&
-      highlightedSuggestion.value &&
-      command.value.trim() !== highlightedSuggestion.value.command
-    ) {
-      event.preventDefault();
-      selectCommand(highlightedSuggestion.value.command);
-      return;
-    }
-    const movedToPlaceholder = selectNextPlaceholder(inputElement.value?.selectionEnd ?? 0);
-    if (!movedToPlaceholder) {
-      return;
-    }
-    event.preventDefault();
-  }
-}
-
-function selectCommand(value: string): void {
-  command.value = value;
-  closeSuggestionMenu();
-  nextTick(() => {
-    const focusedPlaceholder = selectNextPlaceholder(0);
-    if (!focusedPlaceholder) {
-      executeButtonElement.value?.focus();
-    }
-  });
-}
-
-function onCommandInput(): void {
-  isSuggestionMenuVisible.value = hasSuggestions.value;
-  isSuggestionNavigationActive.value = false;
-  resetHighlight();
-}
-
-function onCommandInputFocus(): void {
-  if (suppressSuggestionMenuOnFocus.value) {
-    suppressSuggestionMenuOnFocus.value = false;
-    return;
-  }
-  if (hasSuggestions.value) {
-    isSuggestionMenuVisible.value = true;
-  }
-}
-
-function onCommandInputBlur(): void {
-  window.setTimeout(() => {
-    closeSuggestionMenu();
-  }, 100);
-}
-
 function openCommandHistory(): void {
   isCommandHistoryOpen.value = true;
 }
@@ -256,24 +93,10 @@ function toggleQueueExpansion(): void {
   isQueueExpanded.value = !isQueueExpanded.value;
 }
 
-function onSuggestionActivate(index: number): void {
-  activateSuggestionNavigation(index);
-}
-
-function restoreCommandFromHistory(value: string): void {
-  command.value = value;
+function onSelectCommandFromHistory(value: string): void {
+  restoreCommandFromHistory(value);
   isCommandHistoryOpen.value = false;
-  nextTick(() => inputElement.value?.focus());
 }
-
-watch(settings, value => {
-  const normalized = normalizeSettings(value);
-  if (!areSettingsEqual(value, normalized)) {
-    settings.value = normalized;
-    return;
-  }
-  persistSettings(value);
-}, { deep: true });
 
 watch(() => settings.value.continueIfOneFails, value => {
   updateContinueIfOneFails(value);
@@ -281,10 +104,6 @@ watch(() => settings.value.continueIfOneFails, value => {
 
 watch(() => settings.value.historySize, () => {
   trimLogs();
-});
-
-watch(() => settings.value.comHistorySize, () => {
-  trimToSize();
 });
 
 watch(remoteContinueIfOneFails, value => {
@@ -308,12 +127,6 @@ watch(queueLength, value => {
   if (value < 1) {
     activeTab.value = 'log';
     isQueueExpanded.value = false;
-  }
-});
-
-watch(hasSuggestions, value => {
-  if (!value) {
-    closeSuggestionMenu();
   }
 });
 </script>
@@ -508,7 +321,7 @@ watch(hasSuggestions, value => {
     <CommandHistoryDialog
       v-model="isCommandHistoryOpen"
       :entries="commandHistoryEntries"
-      @select-command="restoreCommandFromHistory"
+      @select-command="onSelectCommandFromHistory"
     />
     <SettingsDialog
       v-model="isSettingsDialogOpen"
