+++ mycore-webcli/src/main/vue/webcli/src/components/CommandHistoryDialog.vue
@@ -1,5 +1,7 @@
 <script setup lang="ts">
-import { nextTick, onBeforeUnmount, ref, watch } from 'vue';
+import { nextTick, ref, toRef } from 'vue';
+
+import { useDialogLifecycle } from '@/composables/useDialogLifecycle';
 
 const props = defineProps<{
   entries: string[];
@@ -12,7 +14,6 @@ const emit = defineEmits<{
 }>();
 
 const dialogRef = ref<HTMLDialogElement | null>(null);
-let previousFocus: HTMLElement | null = null;
 
 function closeDialog(): void {
   emit('update:modelValue', false);
@@ -22,34 +23,10 @@ function selectCommand(value: string): void {
   emit('select-command', value);
 }
 
-watch(() => props.modelValue, async value => {
-  const dialog = dialogRef.value;
-  if (!dialog) {
-    return;
-  }
-  if (value) {
-    previousFocus = document.activeElement instanceof HTMLElement ? document.activeElement : null;
-    if (typeof dialog.showModal === 'function' && !dialog.open) {
-      dialog.showModal();
-    } else {
-      dialog.setAttribute('open', '');
-    }
-    await nextTick();
-    dialog.querySelector<HTMLButtonElement>('.list-group-item, .btn-close')?.focus();
-    return;
-  }
-  if (dialog.open && typeof dialog.close === 'function') {
-    dialog.close();
-  } else {
-    dialog.removeAttribute('open');
-  }
-  previousFocus?.focus();
-});
-
-onBeforeUnmount(() => {
-  if (dialogRef.value?.open && typeof dialogRef.value.close === 'function') {
-    dialogRef.value.close();
-  }
+useDialogLifecycle(toRef(props, 'modelValue'), dialogRef, () => {
+  nextTick(() => {
+    dialogRef.value?.querySelector<HTMLButtonElement>('.list-group-item, .btn-close')?.focus();
+  });
 });
 </script>
